Improve resource clean up. Improve printing.
The resource clean up closes more files and does so in an order that
preserves the parent process' ability to print messages on the console,
instead into the log, for as long as possible.

The printing infrastructure now produces the same output for both the
parallel and single threaded use cases. The order of test cases for the
parallel use case is still non-deterministic, but the output for each
test case is the same.